Added the beginning of two new commands, `scan-organization` and `process-organization`.
The first will scan an organization to gather which repos are configured. It simply reports back on whether the configuration file exists.

The second will scan an organization and pull down all of the configuration files. This will eventually kick off processing for each repository.

Also renamed `scan-repo` to `scan-repository`.